[cow] Teach uniqueness about mark_dependence on the array value.
We already did this for element accesses, this just extends it to handle array value.

This prevents a regression that is caused by my needing to insert an enum
so I can properly RAW a strong_pin operand onto the strong pin's users in a
subsequent patch to improve RemovePin for +0 self.

rdar://20267677

Swift SVN r26528