Make sure we have operands before dereferencing Operands[1] in CondBranchInst::get{True,False}Operands().
Without this if you called either of these methods when you did not have True or
False operands, memory that is not owned by the CondBranchInst would be touched.
Now we just check if we don't have the relevant arguments and early return an
empty array of the relevant type.

Swift SVN r26782