SimplifyCFG: rewrite the dominator based condition simplification.
The optimization simplifies a condition if it is dominated by a terminating instruction with the same condition.
With this change we can also simplify non-terminator instructions, e.g. a cond_fail which is dominated by a cond_br,
or a select_enum which is dominated by a switch_enum.

I do not handle the pattern cond_br(select_enum) anymore, because SILCombine converts this to switch_enum anyway.

There is no notable effect on the benchmarks. But it simplifies the class array code a bit.



Swift SVN r26809