Fix <rdar://problem/20135489> Enum case labels are ignored in "case let" statements
Now that tuple pattern labels are squared away, we can detect mismatches and produce
diagnostics that point exactly to the problematic label.  Right now we require the
tuple labels to match up (or be absent) from the pattern, and produce this diagnostic
if not:

x.swift:7:20: error: tuple pattern element label 'xxx' must be 'one'
    case let .Case(xxx: _, two: x):
                   ^

In the future we could support reshuffling if it someone cares enough to support it.
The important thing for now is to fix the miscompilation.





Swift SVN r26865