Remove invalidation of loop analysis.
We already invalidate all the analyses for each function we inline into,
so this shouldn't be necessary.

We should also be able to remove the invalidation of dominators for the
same reason, but I am getting one test failure when I do that so it
needs further investigation.

Swift SVN r26939