SimplifyCFG: propagateSwitchEnumCondition should not change the CFG
It is called from dominatorBasedSimplify which relies on a correct dominator
tree. Instead when we know the branch target of a switch_enum because of a
dominating switch_enum on the same operand change the dominated switch_enum to
use a enum instruction of the known enum case.

Disable some tests that now fail. They will be enabled again in follow up
commits.

Swift SVN r26975