Update the call graph with newly linked-in functions.
We claim to maintain the call graph in these passes, so we really need
to add nodes for new functions we pull in.

Also, link in functions when building the call graph, and only allow
functions with bodies to be added to the call graph.

This makes the call graph more consistent.

At some point we need to revisit our linking story because we've got
code spread out over several phases now where it might make sense to do
a single up-front linking pass that potentially pulls in
never-referenced functions (e.g. pull in all foo() that could be reached
in a given class hierarchy up front, even if in reality only C.foo() is
ever called).

Swift SVN r27096